+++ app/cdap/components/GridTable/mock/metrics.ts
@@ -1,0 +1,158 @@
+export default [
+  {
+    name: 'customerName',
+    values: [
+      {
+        label: 'Missing/Null',
+        count: 794,
+      },
+      {
+        label: 'Invalid',
+        count: 142,
+      },
+    ],
+  },
+  {
+    name: 'region',
+    values: [
+      {
+        label: 'Missing/Null',
+        count: 456,
+      },
+      {
+        label: 'Invalid',
+        count: 149,
+      },
+    ],
+  },
+  {
+    name: 'creditCardNumber',
+    values: [
+      {
+        label: '8157-1299-9182-9991',
+        count: 191,
+      },
+      {
+        label: '7662-9182-1921-0281',
+        count: 164,
+      },
+    ],
+  },
+  {
+    name: 'purchaseDate',
+    values: [
+      {
+        label: '12-04-2022',
+        count: 794,
+      },
+      {
+        label: '19-06-2022',
+        count: 142,
+      },
+    ],
+  },
+  {
+    name: 'orderedAmount',
+    values: [
+      {
+        label: 'Missing/Null',
+        count: 794,
+      },
+      {
+        label: 'Invalid',
+        count: 142,
+      },
+    ],
+  },
+  {
+    name: 'commissionPercentage',
+    values: [
+      {
+        label: 'Missing/Null',
+        count: 794,
+      },
+      {
+        label: 'Invalid',
+        count: 142,
+      },
+    ],
+  },
+  {
+    name: 'commissionPercentage',
+    values: [
+      {
+        label: 'Missing/Null',
+        count: 794,
+      },
+      {
+        label: 'Invalid',
+        count: 142,
+      },
+    ],
+  },
+  {
+    name: 'commissionPercentage',
+    values: [
+      {
+        label: 'Missing/Null',
+        count: 794,
+      },
+      {
+        label: 'Invalid',
+        count: 142,
+      },
+    ],
+  },
+  {
+    name: 'commissionPercentage',
+    values: [
+      {
+        label: 'Missing/Null',
+        count: 794,
+      },
+      {
+        label: 'Invalid',
+        count: 142,
+      },
+    ],
+  },
+  {
+    name: 'commissionPercentage',
+    values: [
+      {
+        label: 'Missing/Null',
+        count: 794,
+      },
+      {
+        label: 'Invalid',
+        count: 142,
+      },
+    ],
+  },
+  {
+    name: 'commissionPercentage',
+    values: [
+      {
+        label: 'Missing/Null',
+        count: 794,
+      },
+      {
+        label: 'Invalid',
+        count: 142,
+      },
+    ],
+  },
+  {
+    name: 'commissionPercentage',
+    values: [
+      {
+        label: 'Missing/Null',
+        count: 794,
+      },
+      {
+        label: 'Invalid',
+        count: 142,
+      },
+    ],
+  },
+];
